reorganized IO modules and reflect changes in weights and downtimes connector
