+++ modules/io/avrowrite.py
@@ -1,0 +1,51 @@
+import avro.schema
+from avro.datafile import DataFileWriter
+from avro.io import DatumWriter, BinaryEncoder
+
+from io import BytesIO
+
+class AvroWriteException(BaseException):
+    pass
+
+def load_schema(schema):
+    try:
+        f = open(schema)
+        schema = avro.schema.parse(f.read())
+        return schema
+    except Exception as e:
+        raise e
+
+class AvroWriter(object):
+    """ AvroWriter """
+    def __init__(self, schema, outfile):
+        self.schema = schema
+        self.outfile = outfile
+        self.datawrite = None
+        self.avrofile = None
+        self._load_datawriter()
+
+    def _load_datawriter(self):
+        try:
+            lschema = load_schema(self.schema)
+            self.avrofile = open(self.outfile, 'w+b')
+            self.datawrite = DataFileWriter(self.avrofile, DatumWriter(), lschema)
+        except Exception:
+            return False
+
+        return True
+
+    def write(self, data):
+        try:
+            if (not self.datawrite or not self.avrofile):
+                raise AvroWriteException('AvroFileWriter not initalized')
+
+            for elem in data:
+                self.datawrite.append(elem)
+
+            self.datawrite.close()
+            self.avrofile.close()
+
+        except Exception as e:
+            return False, e
+
+        return True, None
