+++ modules/io/statewrite.py
@@ -1,0 +1,41 @@
+import datetime
+import os
+import json
+import requests
+
+from argo_egi_connectors.helpers import datestamp, retry, module_class_name
+
+
+daysback = 1
+
+
+def state_write(caller, statedir, state, savedays, date=None):
+    filenamenew = ''
+    if 'topology' in caller:
+        filenamebase = 'topology-ok'
+    elif 'metricprofile' in caller:
+        filenamebase = 'metricprofile-ok'
+    elif 'weights' in caller:
+        filenamebase = 'weights-ok'
+    elif 'downtimes' in caller:
+        filenamebase = 'downtimes-ok'
+
+    if date:
+        datebackstamp = date
+    else:
+        datebackstamp = datestamp(daysback)
+
+    filenamenew = filenamebase + '_' + datebackstamp
+    db = datetime.datetime.strptime(datebackstamp, '%Y_%m_%d')
+
+    datestart = db - datetime.timedelta(days=int(savedays))
+    i = 0
+    while i < int(savedays) * 2:
+        d = datestart - datetime.timedelta(days=i)
+        filenameold = filenamebase + '_' + d.strftime('%Y_%m_%d')
+        if os.path.exists(statedir + '/' + filenameold):
+            os.remove(statedir + '/' + filenameold)
+        i += 1
+
+    with open(statedir + '/' + filenamenew, 'w') as fp:
+        fp.write(str(state))
